CCS-35 fix(s3): 리스트가 비어 있을 경우 S3 요청이 가지 않도록 수정

+++ backend/src/main/java/com/trend_now/backend/image/application/ImagesService.java
@@ -78,12 +78,14 @@ public class ImagesService {
      */
     @Transactional
     public void deleteImageByPostId(Long postId) {
-        // S3에서 이미지 삭제
         List<String> s3Keys = imagesRepository.findS3KeyByPostsId(postId);
-        s3Service.deleteByS3KeyList(s3Keys);
+        if (s3Keys != null && !s3Keys.isEmpty()) {
+            // S3에서 이미지 삭제
+            s3Service.deleteByS3KeyList(s3Keys);
+            // DB에서 이미지 삭제
+            imagesRepository.deleteAllByPosts_Id(postId);
+        }
 
-        // DB에서 이미지 삭제
-        imagesRepository.deleteAllByPosts_Id(postId);
     }
